requirements.txt
absl-py==2.3.1
accessible-pygments==0.0.5
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alabaster==1.0.0
alembic==1.17.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
autopage==0.5.2
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
celluloid==0.2.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
cliff==4.11.0
cmaes==0.12.0
cmd2==2.7.0
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
coverage==7.10.7
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
docutils==0.21.2
einops==0.7.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
gnureadline==8.2.13
grpcio==1.75.1
gudhi==3.11.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
hypernetx==1.2.5
identify==2.6.15
idna==3.11
igraph==0.11.9
imagesize==1.4.1
iniconfig==2.1.0
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.3.0
lightning==2.4.0
lightning-utilities==0.15.2
littleutils==0.2.4
Mako==1.3.10
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
myst-parser==4.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbsphinx==0.9.7
nbsphinx-link==1.3.1
nest-asyncio==1.6.0
networkx==2.8.8
nodeenv==1.9.1
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.26.4
numpydoc==1.9.0
ogb==1.3.6
omegaconf==2.3.0
optuna==2.10.1
outdated==0.2.2
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
prettytable==3.16.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydata-sphinx-theme==0.16.1
pyg-nightly==2.7.0.dev20251015
Pygments==2.19.2
pyparsing==3.2.5
pyperclip==1.11.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-igraph==0.11.9
python-json-logger==4.0.0
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-argparse==1.7.1
rootutils==1.0.7
rpds-py==0.27.1
ruff==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
Send2Trash==1.8.3
sentry-sdk==2.42.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.8
SpharaPy==1.1.2
Sphinx==8.1.3
sphinx-gallery==0.19.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.44
stack-data==0.6.3
stevedore==5.5.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tinycss2==1.4.0
-e git+https://github.com/geometric-intelligence/TopoBench@c5216f372a0389d89a596e31588ca46372a35db2#egg=TopoBench
TopoModelX @ git+https://github.com/pyt-team/TopoModelX.git@8fc047c85886642cf53a784fc712cbae89b015fd
TopoNetX @ git+https://github.com/pyt-team/TopoNetX.git@39ed2c52fba9faaa0c499a51282d91e72153630e
torch==2.3.0
torch_cluster==1.6.3
torch_geometric==2.5.2
torch_scatter==2.1.2
torch_sparse==0.6.18
torch_spline_conv==1.2.2
torchmetrics==1.8.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.8.3
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
virtualenv==20.35.3
wandb==0.22.2
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xxhash==3.6.0
yarl==1.22.0
